requirements.txt
langchain
langchain-google-genai
langchain-community
google-generativeai
openai
pandas
tqdm
regex
matplotlib
